% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{listPlotWidgets}
\alias{listPlotWidgets}
\title{List available plot widgets}
\usage{
listPlotWidgets()
}
\value{
Invisibly returns a named list of widget functions
}
\description{
List available plot widgets
}
\details{
Simply print out all available wigdets. A list containing the widget
function is returned invisibly.
}
\examples{
listPlotWidgets()
}

